pyproject.toml
[tool.poetry]
name = "labs-auth"
version = "0.1.0"
description = "An authentication system"
authors = ["Weston Gorczyca <weston.gorczyca@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
waitress = "^3.0.0"
flask-login = "^0.6.3"
flask-redis = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
